# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: AfuCloud API – cloud platform for storing, managing, and delivering digital assets
servers:
  - url: /api
    description: Base API path
security:
  - BearerAuth: []
...
paths:
...
  /v1/projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /v1/projects/{projectId}/images:
    get:
      operationId: listImages
      tags: [images]
      summary: List images in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: album
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
        - name: favorite
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Image list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageListResponse"
  /v1/projects/{projectId}/images/{id}:
    get:
      operationId: getImage
      tags: [images]
      summary: Get image details
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
    patch:
      operationId: updateImage
      tags: [images]
      summary: Update image metadata
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageUpdate"
      responses:
        "200":
          description: Image updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
    delete:
      operationId: deleteImage
      tags: [images]
      summary: Delete (soft-delete) an image
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/projects/{projectId}/images/{id}/favorite:
    patch:
      operationId: toggleImageFavorite
      tags: [images]
      summary: Toggle image favorite status
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Favorite toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
  /v1/projects/{projectId}/images/{id}/restore:
    post:
      operationId: restoreImage
      tags: [images]
      summary: Restore a soft-deleted image
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"
  /v1/projects/{projectId}/images/upload-url:
    post:
      operationId: getUploadUrl
      tags: [images]
      summary: Get a pre-signed URL for direct upload
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Pre-signed upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
  /v1/projects/{projectId}/images/confirm-upload:
    post:
      operationId: confirmUpload
      tags: [images]
      summary: Confirm a completed direct upload and persist image record
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmUploadInput"
      responses:
        "201":
          description: Image record confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Image"

  # ─── API KEYS ───
  /v1/projects/{projectId}/api-keys:
    get:
      operationId: listApiKeys
      tags: [api-keys]
      summary: List API keys for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
    post:
      operationId: createApiKey
      tags: [api-keys]
      summary: Create a new API key
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKeyInput"
      responses:
        "201":
          description: API key created (secret shown once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyWithSecret"
  /v1/projects/{projectId}/api-keys/{id}:
    delete:
      operationId: revokeApiKey
      tags: [api-keys]
      summary: Revoke an API key
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Key revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── PERSONAL ACCESS TOKENS ───
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: AfuCloudCredential
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── Auth ───
...
    Project:
      type: object
      required: [id, name, slug, userId, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        userId:
          type: string
        imageCount:
          type: number
        storageUsed:
          type: number
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
...
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
...
    Image:
      type: object
      required: [id, projectId, name, url, format, size, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        originalName:
          type: ["string", "null"]
        url:
          type: string
        publicUrl:
          type: ["string", "null"]
        downloadUrl:
          type: string
        format:
          type: string
        size:
          type: number
        width:
          type: ["number", "null"]
        height:
          type: ["number", "null"]
        favorite:
          type: boolean
        tags:
          type: array
          items:
            type: string
        album:
          type: ["string", "null"]
        deletedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ImageListResponse:
      type: object
      required: [images, total, page, limit]
      properties:
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
    ImageUpdate:
      type: object
      properties:
        name:
          type: string
        tags:
          type: array
          items:
            type: string
        album:
          type: ["string", "null"]
    UploadUrlInput:
      type: object
      required: [filename, contentType]
      properties:
        filename:
          type: string
        contentType:
          type: string
        name:
          type: string
    UploadUrlResponse:
      type: object
      required: [uploadUrl, imageId, key]
      properties:
        uploadUrl:
          type: string
        imageId:
          type: string
        key:
          type: string
    ConfirmUploadInput:
      type: object
      required: [imageId, key]
      properties:
        imageId:
          type: string
        key:
          type: string
        width:
          type: number
        height:
          type: number
        size:
          type: number
        tags:
          type: array
          items:
            type: string
        album:
          type: string

    # ─── API Keys ───
    ApiKey:
      type: object
      required: [id, projectId, name, environment, prefix, createdAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        environment:
          type: string
        prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        lastUsedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ApiKeyInput:
      type: object
      required: [name, environment]
      properties:
        name:
          type: string
        environment:
          type: string
        scopes:
          type: array
          items:
            type: string
    ApiKeyWithSecret:
      allOf:
        - $ref: "#/components/schemas/ApiKey"
        - type: object
          required: [secret]
          properties:
            secret:
              type: string

    # ─── Tokens ───
...